Add an optional "device" parameter the the isn() functions
I basically tried to add the cuda parameter for using the GPU but I have
no dear clue if this will ever work; I also opened an issue some time
ago, dunno if it's still there, but I already has listed in there some
problems I was facing with the library, e.g. we should be able to the
isn-computation functions whatever metric we are pleased with, but that
was not working by the time I tried it; also the metrics need to be
changed somehow - how I wanna keep all of them tbh, I just want to have
as "fixed" metrics the fastest ones but then I wanna keep also the
others to be able to choose (this makes more sense in my brain than
written all out here). Once all this is proved to be working (and
hopefully very very very soon), I can move on to the benchmark (FINALLY)
and then to finish that awfully stupid paper and shall my life be
compelte? Who knows.

Co-Authored-By: James Collier <james.collier@vib.be>